0.8.0のリリース

+++ lib/rossoc/backend.rb
@@ -10,10 +10,11 @@ module Rossoc
 
     RESERVED_TABLES = %w[mruby arduino dev].freeze
 
-    def initialize(ir, output)
+    def initialize(ir, output, overwrite_enable)
       @ir = ir
       @output = output
       @content = nil
+      @overwrite_enable = overwrite_enable
     end
 
     def generate
@@ -32,6 +33,8 @@ module Rossoc
       raise BackendError, 'No content' if @content.nil?
 
       raise BackendError, 'No output file' if @output.blank?
+
+      raise BackendError, "File exists #{@output}" if File.exist?(@output) && !@overwrite_enable
 
       file = File.open(@output, 'w')
       file.write(@content)
